Feature/AI: first API commit
